% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble-phyloseq}
\alias{as_tibble-phyloseq}
\alias{as_tibble.otu_table}
\alias{as_tibble.sample_data}
\alias{as_tibble.taxonomyTable}
\alias{as_tibble.XStringSet}
\alias{as_tibble.phyloseq}
\title{Coerce phyloseq objects to tibble data frames}
\usage{
\method{as_tibble}{otu_table}(x, pivot = TRUE, .name_repair = base::make.unique)

\method{as_tibble}{sample_data}(x, .name_repair = base::make.unique)

\method{as_tibble}{taxonomyTable}(x, .name_repair = base::make.unique)

\method{as_tibble}{XStringSet}(x)

\method{as_tibble}{phyloseq}(x, tax = TRUE, ref = FALSE, .name_repair = base::make.unique)
}
\arguments{
\item{x}{A phyloseq object or component.}

\item{pivot}{Whether to pivot the otu table to long format.}

\item{.name_repair}{Function to repair names in the case of conflicts.}

\item{tax}{Whether to include taxonomy data.}

\item{ref}{Whether to include reference sequences.}
}
\value{
A tibble (tbl_df)
}
\description{
These functions extend the \code{as_tibble()} function defined by the tibble
package to work on phyloseq objects or phyloseq-component objects.
}
\details{
Tibbles (tbl_df objects) do not support rownames; the taxa and sample names
in the returned tibbles will always be stored in the first or second
columns. The names ".otu", ".sample", ".abundance", and ".sequence" are
special column names reserved for the otu/taxa names, sample names,
abundances, and reference sequences.
}
\examples{
library(tibble) # for as_tibble() and glimpse()

data(GlobalPatterns)

# Subset to 1/100 of the original taxa to speed operations
ps <- GlobalPatterns \%>\%
  filter_tax_table(dplyr::row_number() \%\% 100 == 1)

# On phyloseq objects, as_tibble is similar to psmelt()
psmelt(ps) \%>\% glimpse
as_tibble(ps) \%>\% glimpse

# By default, the otu_table method provides a tibble in long-format like
# psmelt and the phyloseq method
otu_table(ps) \%>\% as_tibble \%>\% glimpse

# Sample data and taxonomy tables produced by as_tibble can be converted
# back into their respective phyloseq objects using speedyseq's tbl_df
# constructors
sample_data(ps) <- sample_data(ps) \%>\%
  as_tibble \%>\%
  dplyr::mutate(sample_sum = sample_sums(ps)) \%>\%
  sample_data
sample_data(ps) \%>\% glimpse
}
