% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble-speedyseq}
\alias{as_tibble-speedyseq}
\alias{as_tibble.otu_table}
\alias{as_tibble.sample_data}
\alias{as_tibble.taxonomyTable}
\alias{as_tibble.XStringSet}
\alias{as_tibble.phyloseq}
\title{Coerce phyloseq objects to tibbles (tbl_df objects)}
\usage{
as_tibble.otu_table(x, pivot = TRUE, .name_repair = base::make.unique)

as_tibble.sample_data(x, .name_repair = base::make.unique)

as_tibble.taxonomyTable(x, .name_repair = base::make.unique)

as_tibble.XStringSet(x)

as_tibble.phyloseq(
  x,
  tax = TRUE,
  ref = FALSE,
  .name_repair = base::make.unique
)
}
\arguments{
\item{x}{a phyloseq object or component}

\item{pivot}{whether to pivot the data frame to long format}

\item{tax}{whether to include taxonomy data}
}
\description{
These functions extend the \code{as_tibble()} function defined by the tibble
package to work on phyloseq objects or phyloseq-component objects.
}
\section{Functions}{
\itemize{
\item \code{as_tibble.otu_table}: method for otu_table objects

\item \code{as_tibble.sample_data}: method for sample_data objects

\item \code{as_tibble.taxonomyTable}: method for taxonomyTable objects

\item \code{as_tibble.XStringSet}: method for DNAStringSet objects

\item \code{as_tibble.phyloseq}: method for phyloseq objects
}}

