% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_taxa_vec.R
\name{bad_flush_right}
\alias{bad_flush_right}
\title{Replace all values with NA upon seeing a bad value}
\usage{
bad_flush_right(x, bad = "BAD", na_bad = FALSE, k = length(x))
}
\arguments{
\item{x}{a vector}

\item{bad}{the string representing a bad value}

\item{na_bad}{whether NAs should also be treated as bad}

\item{k}{the index to which values are flushed}
}
\description{
Helper for \code{merge_taxa_vec()}
}
\keyword{internal}
