% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_taxa_vec.R
\name{bad_or_unique}
\alias{bad_or_unique}
\title{Reduce a vector x to its unique value or the value of \code{bad}}
\usage{
bad_or_unique(x, bad = "BAD")
}
\arguments{
\item{x}{a vector}

\item{bad}{the string representing a bad value}
}
\description{
Helper for \code{merge_taxa_vec()}
}
\keyword{internal}
