% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{filter-phyloseq}
\alias{filter-phyloseq}
\alias{filter_tax_table}
\alias{filter_tax_table,phyloseq-method}
\alias{filter_tax_table,taxonomyTable-method}
\alias{filter_sample_data}
\alias{filter_sample_data,phyloseq-method}
\alias{filter_sample_data,sample_data-method}
\title{Subset rows in the taxonomy table or sample data using column values}
\usage{
filter_tax_table(x, ...)

\S4method{filter_tax_table}{phyloseq}(x, ...)

\S4method{filter_tax_table}{taxonomyTable}(x, ...)

filter_sample_data(x, ...)

\S4method{filter_sample_data}{phyloseq}(x, ...)

\S4method{filter_sample_data}{sample_data}(x, ...)
}
\arguments{
\item{x}{A \code{phyloseq}, \code{taxonomyTable}, or \code{sample_data} object}

\item{...}{Expressions passed to \code{dplyr::filter()}}
}
\description{
These functions are wrappers around \code{dplyr::filter()} that make it possible
to subset rows (corresponding to taxa or samples) using column values in the
taxonomy table or sample data.
}
\examples{
data(GlobalPatterns)

ps <- GlobalPatterns \%>\%
  filter_tax_table(Kingdom == "Bacteria") \%>\%
  filter_sample_data(SampleType \%in\% c("Feces", "Soil"))
}
