% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{filter_sample_data}
\alias{filter_sample_data}
\alias{filter_sample_data,phyloseq-method}
\alias{filter_sample_data,sample_data-method}
\title{Subset samples using values in the sample data}
\usage{
filter_sample_data(x, ...)

\S4method{filter_sample_data}{phyloseq}(x, ...)

\S4method{filter_sample_data}{sample_data}(x, ...)
}
\arguments{
\item{x}{A \code{phyloseq} or \code{sample_data} object}

\item{...}{Expressions passed to \code{dplyr::filter()}}
}
\description{
This function is a wrapper around \code{dplyr::filter()} that provides a
convenient way to subset samples using the \code{sample_data(x)}.
}
