% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{filter_tax_table}
\alias{filter_tax_table}
\alias{filter_tax_table,phyloseq-method}
\alias{filter_tax_table,taxonomyTable-method}
\title{Subset taxa using values in the taxonomy table}
\usage{
filter_tax_table(x, ...)

\S4method{filter_tax_table}{phyloseq}(x, ...)

\S4method{filter_tax_table}{taxonomyTable}(x, ...)
}
\arguments{
\item{x}{A \code{phyloseq} or \code{taxonomyTable} object}

\item{...}{Expressions passed to \code{dplyr::filter()}}
}
\description{
This function makes it possible to choose a subset of taxa using values in
the taxonomy table as in \code{dplyr::filter()}. The taxonomy table is converted
to a tibble using \code{ps_tibble()} and the arguments in \code{...} are passed
directly to \code{dplyr::filter()}. The taxonomy table in the phyloseq object is
then updated to contain just the subset taxa.
}
