% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_samples.R
\name{merge_groups}
\alias{merge_groups}
\title{Merge groups of elements within a vector by a function}
\usage{
merge_groups(x, group, f = unique_or_na)
}
\arguments{
\item{x}{A vector whose elements will be merged.}

\item{group}{A vector such that \code{as.factor(group)} defines the grouping.}

\item{f}{A function that, when applied to a subvector of x, returns a single
value. Can also be a formula as interpretted by \code{purrr::as_mapper()}.}
}
\description{
Internal function used in \code{merge_samples2()} to merge variables. Note, owing
to the use of \code{split()}, the merged elements in the new vector will be
reordered according to \code{group}.
}
\keyword{internal}
