% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_samples.R
\name{merge_samples2}
\alias{merge_samples2}
\alias{merge_samples2,phyloseq-method}
\alias{merge_samples2,otu_table-method}
\alias{merge_samples2,sample_data-method}
\title{Merge samples by sample variables}
\usage{
merge_samples2(x, group, funs = list(), reorder = FALSE)

\S4method{merge_samples2}{phyloseq}(x, group, funs = list(), reorder = FALSE)

\S4method{merge_samples2}{otu_table}(x, group, reorder = FALSE)

\S4method{merge_samples2}{sample_data}(x, group, funs = list(), reorder = FALSE)
}
\arguments{
\item{x}{A \code{sample_data} object}

\item{group}{A sample variable or a vector of length \code{nsamples(x)} defining
the sample grouping. A vector must be supplied if x is an otu_table}

\item{funs}{Named list of merge functions for each numeric variable}

\item{reorder}{Logical specifying whether to reorder the new (merged)
samples by name}
}
\description{
Alternative to \code{phyloseq::merge_samples()} that better handles categorical
sample variables. Using the \code{funs} argument, users can specify the function
that should be used to merge sample variables within groups. The default is
use the new \code{unique_or_na()} function, which collapses the values to a
single unique value if it exists and otherwise returns NA.
}
\examples{
data(enterotype)

 # Merge samples with the same project and clinical status
ps <- enterotype 
sample_data(ps) <- sample_data(ps) \%>\%
  transform(Project.ClinicalStatus = Project:ClinicalStatus)
sample_data(ps) \%>\% head
ps0 <- merge_samples2(ps, "Project.ClinicalStatus", funs = list(Age = mean))
sample_data(ps0) \%>\% head
}
