% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_taxa_vec.R
\name{merge_taxa_vec_pseudo}
\alias{merge_taxa_vec_pseudo}
\title{Pseudo taxa merging for phylo and XStringSet objects}
\usage{
merge_taxa_vec_pseudo(x, group)
}
\arguments{
\item{x}{a phylo or XStringSet object}

\item{group}{a vector with one element for each taxon in `physeq` that
defines the new groups}
}
\description{
Returns `x` pruned to the first taxon of each group defined in `group`.
}
\keyword{internal}
