% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{mutate_sample_data}
\alias{mutate_sample_data}
\alias{mutate_sample_data,phyloseq-method}
\alias{mutate_sample_data,sample_data-method}
\title{Create, modify, and delete columns in the sample data}
\usage{
mutate_sample_data(x, ...)

\S4method{mutate_sample_data}{phyloseq}(x, ...)

\S4method{mutate_sample_data}{sample_data}(x, ...)
}
\arguments{
\item{x}{A \code{phyloseq} or \code{sample_data} object}

\item{...}{Expressions passed to \code{dplyr::mutate()}}
}
\description{
This function is a wrapper around \code{dplyr::mutate()} that provides a
convenient way to modify \code{sample_data(x)}.
}
