% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{mutate_tax_table}
\alias{mutate_tax_table}
\alias{mutate_tax_table,phyloseq-method}
\alias{mutate_tax_table,taxonomyTable-method}
\title{Create, modify, and delete columns in the taxonomy table}
\usage{
mutate_tax_table(x, ...)

\S4method{mutate_tax_table}{phyloseq}(x, ...)

\S4method{mutate_tax_table}{taxonomyTable}(x, ...)
}
\arguments{
\item{x}{A \code{phyloseq} or \code{taxonomyTable} object}

\item{...}{Expressions passed to \code{dplyr::mutate()}}
}
\description{
This function is a wrapper around \code{dplyr::mutate()} that provides a
convenient way to modify \code{tax_table(x)}.
The \code{.otu} column name can be used to set new taxa names.
}
