% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{oneline}
\alias{oneline}
\title{One-line summaries of phyloseq components}
\usage{
oneline(x)
}
\arguments{
\item{x}{A phyloseq component object}
}
\description{
One-line summaries of phyloseq components
}
\keyword{internal}
