% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orient_taxa}
\alias{orient_taxa}
\title{Orient a phyloseq object or otu table to have taxa as rows or as columns}
\usage{
orient_taxa(x, as)
}
\arguments{
\item{x}{A phyloseq or otu-table object}

\item{as}{The matrix dimension that is desired for taxa. Must be
"rows" for rows and "columns" or "cols" for columns.}
}
\description{
Puts the phyloseq or otu-table object \code{x} in the orientation (taxa as rows
or as columns) specified by \code{as}. This is useful when passing the otu table
on to functions that require the abundance matrix to have a specific
orientation and are unaware of the \code{taxa_are_rows(x)} property.
}
\examples{
data(soilrep)
taxa_are_rows(soilrep)
x <- soilrep \%>\% orient_taxa(as = "columns")
taxa_are_rows(x)
}
