% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{relocate-phyloseq}
\alias{relocate-phyloseq}
\alias{relocate_tax_table}
\alias{relocate_tax_table,taxonomyTable-method}
\alias{relocate_tax_table,phyloseq-method}
\alias{relocate_sample_data}
\alias{relocate_sample_data,sample_data-method}
\alias{relocate_sample_data,phyloseq-method}
\title{Change column order in the taxonomy table or sample data}
\usage{
relocate_tax_table(x, ...)

\S4method{relocate_tax_table}{taxonomyTable}(x, ...)

\S4method{relocate_tax_table}{phyloseq}(x, ...)

relocate_sample_data(x, ...)

\S4method{relocate_sample_data}{sample_data}(x, ...)

\S4method{relocate_sample_data}{phyloseq}(x, ...)
}
\arguments{
\item{x}{A \code{phyloseq}, \code{taxonomyTable}, or \code{sample_data} object}

\item{...}{Expressions and arguments passed to \code{dplyr::relocate()}}
}
\description{
These functions are wrappers around \code{dplyr::relocate()} that provide
convenient ways to modify \code{tax_table(x)} and \code{sample_data(x)}.
}
\details{
The special column names '.otu' and '.sample' should not be used; see
\code{mutate-phyloseq} for the ability to change taxa and sample names using
these names.
}
\examples{
data(GlobalPatterns)

GlobalPatterns \%>\% sample_variables
ps <- GlobalPatterns \%>\%
  relocate_sample_data(SampleType, .after = X.SampleID)
ps \%>\% sample_variables
}
