% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{rename-phyloseq}
\alias{rename-phyloseq}
\alias{rename_tax_table}
\alias{rename_tax_table,taxonomyTable-method}
\alias{rename_tax_table,phyloseq-method}
\alias{rename_with_tax_table}
\alias{rename_with_tax_table,taxonomyTable-method}
\alias{rename_with_tax_table,phyloseq-method}
\alias{rename_sample_data}
\alias{rename_sample_data,sample_data-method}
\alias{rename_sample_data,phyloseq-method}
\alias{rename_with_sample_data}
\alias{rename_with_sample_data,sample_data-method}
\alias{rename_with_sample_data,phyloseq-method}
\title{Rename columns in the taxonomy table or sample data}
\usage{
rename_tax_table(x, ...)

\S4method{rename_tax_table}{taxonomyTable}(x, ...)

\S4method{rename_tax_table}{phyloseq}(x, ...)

rename_with_tax_table(x, ...)

\S4method{rename_with_tax_table}{taxonomyTable}(x, ...)

\S4method{rename_with_tax_table}{phyloseq}(x, ...)

rename_sample_data(x, ...)

\S4method{rename_sample_data}{sample_data}(x, ...)

\S4method{rename_sample_data}{phyloseq}(x, ...)

rename_with_sample_data(x, ...)

\S4method{rename_with_sample_data}{sample_data}(x, ...)

\S4method{rename_with_sample_data}{phyloseq}(x, ...)
}
\arguments{
\item{x}{A \code{phyloseq}, \code{taxonomyTable}, or \code{sample_data} object}

\item{...}{Renaming expressions passed to \code{dplyr::rename()} or arguments
passed to \code{dplyr::rename_with()}}
}
\description{
\code{rename_tax_table()} and \code{rename_with_tax_table()} provide the functionality
of \code{dplyr::rename()} and \code{dplyr::rename_with()} to phyloseq taxonomy tables;
\code{rename_sample_data()} and \code{rename_with_sample_data()} provide this
functionality to phyloseq sample data.
}
\examples{
GlobalPatterns \%>\% rank_names

ps1 <- GlobalPatterns \%>\%
  rename_tax_table(Domain = Kingdom)
ps1 \%>\% rank_names
ps2 <- ps1 \%>\%
  rename_with_tax_table(stringr::str_to_lower)
ps2 \%>\% rank_names

GlobalPatterns \%>\% sample_variables
ps3 <- GlobalPatterns \%>\%
  rename_with_sample_data(janitor::make_clean_names) \%>\%
  rename_sample_data(sample_id = x_sample_id)
ps3 \%>\% sample_variables
}
