% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{show,phyloseq-method}
\alias{show,phyloseq-method}
\alias{show,taxonomyTable-method}
\alias{show,sample_data-method}
\alias{show,otu_table-method}
\title{Methods for \code{show()} for phyloseq objects.}
\usage{
\S4method{show}{phyloseq}(object)

\S4method{show}{taxonomyTable}(object)

\S4method{show}{sample_data}(object)

\S4method{show}{otu_table}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
See \code{\link[methods]{show}}. Speedyseq provides \code{show()} methods that
overwrite phyloseq's methods for the classes defined in phyloseq.
}
\examples{
data(GlobalPatterns)
show(GlobalPatterns)
GlobalPatterns
}
\seealso{
\code{\link[methods]{show}}
}
