% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq-constructors.R
\name{tibble-constructors}
\alias{tibble-constructors}
\alias{otu_table,tbl_df-method}
\alias{sample_data,tbl_df-method}
\alias{tax_table,tbl_df-method}
\title{Construct phyloseq objects from tibbles}
\usage{
\S4method{otu_table}{tbl_df}(object, taxa_are_rows)

\S4method{sample_data}{tbl_df}(object)

\S4method{tax_table}{tbl_df}(object)
}
\arguments{
\item{object}{A tibble whose first column contains the sample or taxa ids}

\item{taxa_are_rows}{Logical; \code{TRUE} if rows correspond to taxa and \code{FALSE}
if rows correspond to samples}
}
\description{
These methods extend phyloseq's constructor functions to construct phyloseq
components from tibbles (objects with class "tbl_df").
}
\details{
Since tibbles cannot have row names, the sample or taxon identifiers must be
contained in a regular column. Speedyseq currently always uses the first
column for the identifiers that would normally be taken from the row names
by phyloseq's built-in constructors. Thus the first column is assumed to
contain the sample names for \code{sample_data()} and the OTU/taxa names for
\code{tax_table()}; for \code{otu_table()}, the first column is assumed to contain the
sample names if \code{taxa_are_rows = TRUE} and the taxa names if \code{taxa_are_rows = FALSE}, with the other identifier being taken from the remaining column
names.
}
\examples{
\dontrun{
# Read a .csv file with readr, which creates an object of class `tbl_df`
tbl <- readr::read_csv("path/to/otu_table.csv")
# Inspect and check if taxa are rows and that the first column contains the
# sample names or the taxa/OTU names
head(tbl) 
# Create a phyloseq `otu_table` object
otu <- otu_table(tbl, taxa_are_rows = FALSE)

# Read a .csv file with readr, which creates an object of class `tbl_df`
tbl <- readr::read_csv("path/to/sample_data.csv")
# Inspect and check that the first column contains the sample names
head(tbl) 
# Create a phyloseq `sample_data` object
sam <- sample_data(tbl)
}
}
\seealso{
\code{\link[tibble:tbl_df-class]{tibble::tbl_df}}
\code{\link[phyloseq:otu_table-methods]{phyloseq::otu_table()}}
\code{\link[phyloseq:sample_data-methods]{phyloseq::sample_data()}}
\code{\link[phyloseq:tax_table-methods]{phyloseq::tax_table()}}
}
