% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{tibble_print}
\alias{tibble_print}
\title{Print a 2-d data.frame or matrix like a tibble}
\usage{
tibble_print(
  x,
  n = NULL,
  width = NULL,
  n_extra = NULL,
  types = TRUE,
  rows = "rows",
  cols = "variables"
)
}
\arguments{
\item{x}{A data frame or matrix}

\item{n}{Number of rows to print}

\item{width}{Display width}

\item{n_extra}{Number of extra columns after width to summarize after main
display}

\item{types}{Whether to print the second row with the variable types}

\item{rows}{What rows are called in the end summary}

\item{cols}{What columns are called in the end summary}
}
\description{
Print a 2-d data.frame or matrix like a tibble
}
\keyword{internal}
