% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agglomeration.R
\name{tree_glom}
\alias{tree_glom}
\title{Agglomerate taxa to a given phylogenetic resolution.}
\usage{
tree_glom(physeq, resolution, shorten = TRUE,
  criterion = "max_tip_depth")
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}}.}

\item{resolution}{Phylogenetic resolution at which to collapse the tree}

\item{shorten}{Whether to shorten the terminal branches to the collapsed
nodes}

\item{criterion}{Criterion to determine whether a node is collapsed}
}
\value{
A taxonomically-agglomerated, optionally-pruned, object with class matching
the class of \code{physeq}.
}
\description{
This method merges taxa that are sufficiently closely related according to
`phy_tree(physeq)`. It uses
\code{\link{castor::collapse_tree_at_resolution}} to determine the groups of
taxa that are to be aggregated and to produce a new tree with each group
collapsed to a single tip. The new taxa are named according to the most
abundant taxon of each group. The reference sequences and taxonomy table
reflect these "archetypal" taxa.
}
\details{
This method is similar to `phyloseq::tip_glom()` except that it uses the
tree directly, rather than as the source of a distance matrix that is used
for hierarchical clustering. The clustering step of `tip_glom()` is slow and
memory intensive when `ntaxa(physeq)` is large. Since clustering does not
need to be performed, `tree_glom()` is much more efficient.
}
\examples{
data(GlobalPatterns)
ntaxa(GlobalPatterns)
ps <- tree_glom(GlobalPatterns, 0.03)
ntaxa(ps)

}
\seealso{
\code{\link{castor::collapse_tree_at_resolution}} for more information about
tree-collapsing method and parameters 

\code{\link{phyloseq::tip_glom}}

\code{\link{tax_glom}}
}
