% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-joins.R
\name{join-phyloseq}
\alias{join-phyloseq}
\alias{left_join_tax_table}
\alias{left_join_sample_data}
\alias{inner_join_tax_table}
\alias{inner_join_sample_data}
\title{Mutating joins of the taxonomy table or sample data}
\usage{
left_join_tax_table(x, ...)

left_join_sample_data(x, ...)

inner_join_tax_table(x, ...)

inner_join_sample_data(x, ...)
}
\arguments{
\item{x}{A \code{phyloseq}, \code{taxonomyTable}, or \code{sample_data} object}

\item{...}{Arguments passed to respective dplyr joining operation}
}
\value{
An object of the same type as \code{x}, with added columns
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions are wrappers around dplyr joining operations (see
\code{dplyr::mutate-joins}) whose first (\code{x}) argument is either the taxonomy
table or sample data.
}
\details{
When joining by taxonomy table, the .otu column name can be used to match
taxa names.
When joining by sample data, the .sample column name can be used to match
sample names.
}
\examples{
data(GlobalPatterns)

GlobalPatterns \%>\% sample_variables
ps1 <- GlobalPatterns \%>\%
  select_sample_data(!contains("Barcode"))
y <- GlobalPatterns \%>\%
  sample_data \%>\%
  select_sample_data(contains("Barcode")) \%>\%
  ps_tibble
ps2 <- ps1 \%>\% left_join_sample_data(y, by = ".sample")
ps2 \%>\% sample_variables
}
\seealso{
\code{\link[dplyr:mutate-joins]{dplyr::left_join()}}
\code{\link[dplyr:mutate-joins]{dplyr::inner_join()}}
}
