% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_samples.R
\name{merge_samples2}
\alias{merge_samples2}
\alias{merge_samples2,phyloseq-method}
\alias{merge_samples2,otu_table-method}
\alias{merge_samples2,sample_data-method}
\title{Merge samples by a sample variable or factor}
\usage{
merge_samples2(x, group, funs = list(), reorder = FALSE)

\S4method{merge_samples2}{phyloseq}(x, group, funs = list(), reorder = FALSE)

\S4method{merge_samples2}{otu_table}(x, group, reorder = FALSE)

\S4method{merge_samples2}{sample_data}(x, group, funs = list(), reorder = FALSE)
}
\arguments{
\item{x}{A \code{sample_data} object}

\item{group}{A sample variable or a vector of length \code{nsamples(x)} defining
the sample grouping. A vector must be supplied if x is an otu_table}

\item{funs}{Named list of merge functions for each numeric variable}

\item{reorder}{Logical specifying whether to reorder the new (merged)
samples by name}
}
\description{
This function provides an alternative to \code{phyloseq::merge_samples()} that
better handles sample variables of different types, especially categorical
sample variables. It combines the samples in \code{x} defined by the sample
variable or factor \code{group} by summing the abundances in \code{otu_table(x)} and
combines sample variables by the summary functions in \code{funs}. The default
summary function, \code{unique_or_na()}, collapses the values within a group to a
single unique value if it exists and otherwise returns NA. The new (merged)
samples are named by the values in \code{group}.
}
\examples{
data(enterotype)

# Merge samples with the same project and clinical status
ps <- enterotype 
sample_data(ps) <- sample_data(ps) \%>\%
  transform(Project.ClinicalStatus = Project:ClinicalStatus)
sample_data(ps) \%>\% head
ps0 <- merge_samples2(ps, "Project.ClinicalStatus", funs = list(Age = mean))
sample_data(ps0) \%>\% head
}
