% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_taxa_vec.R
\name{merge_taxa_vec}
\alias{merge_taxa_vec}
\alias{merge_taxa_vec,phyloseq-method}
\alias{merge_taxa_vec,otu_table-method}
\title{Merge taxa in groups (vectorized version)}
\usage{
merge_taxa_vec(x, group, tax_adjust = 1L)

\S4method{merge_taxa_vec}{phyloseq}(x, group, tax_adjust = 1L)

\S4method{merge_taxa_vec}{otu_table}(x, group)
}
\arguments{
\item{x}{A phyloseq object or component object}

\item{group}{A vector with one element for each taxon in `physeq` that
defines the new groups. See `base::rowsum()`.}

\item{tax_adjust}{0: no adjustment; 1: phyloseq-compatible adjustment; 2:
conservative adjustment (see Details)}
}
\description{
Merge taxa in `x` into a smaller set of taxa defined by the vector `group`.
Taxa whose value in `group` is NA will be dropped. New taxa will be named
either by the most abundant taxon in each group (`name == "max"`).
}
\details{
The `tax_adjust` argument controls the handling of taxonomic disagreements
within groups. Setting `tax_adjust == 0` causes no adjustment; the taxonomy
of the new group is set to the archetype taxon (see below). Setting
`tax_adjust == 1` corresponds to the original phyloseq behavior.
Disagreements within a group at any rank cause the values at lower ranks to
be set to `NA`, but ranks where all taxa in the group are already NA are not
treated as disagreements. As a result, taxonomic assignments where e.g. the
family is NA but the genus is not are be propagated to the new table.
Setting `tax_adjust == 2` (the default) goes further, setting all ranks to
NA after the first disagreement or NA.
}
\examples{
\dontrun{
# Create 97\% OTUs from a phyloseq object with reference sequences
# Assume ps is a phyloseq object with refseq slot
dna <- refseq(ps)
nproc <- 1 # Increase to use multiple processors
aln <- DECIPHER::AlignSeqs(dna, processors = nproc)
d <- DECIPHER::DistanceMatrix(aln, processors = nproc)
clusters <- DECIPHER::IdClusters(
  d, 
  method = "complete",
  cutoff = 0.03,
  processors = nproc
)
ps0 <- merge_taxa_vec(
  ps, 
  group = clusters$cluster,
  tax_adjust = 2
)
}
}
\seealso{
\code{\link[phyloseq]{merge_taxa}}

\code{\link{tax_glom}}
}
