% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_taxa_vec.R
\name{merge_taxa_vec_pseudo}
\alias{merge_taxa_vec_pseudo}
\title{Pseudo-merge taxa in groups}
\usage{
merge_taxa_vec_pseudo(x, group, reorder = FALSE)
}
\arguments{
\item{x}{a phyloseq component-class object}

\item{group}{a vector with one element for each taxon in \code{x} that defines
the new groups}
}
\description{
Returns \code{x} pruned to the first taxon of each group defined in \code{group}.
}
\keyword{internal}
