% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_tibble.R
\name{ps_tibble}
\alias{ps_tibble}
\alias{ps_tibble,otu_table-method}
\alias{ps_tibble,sample_data-method}
\alias{ps_tibble,taxonomyTable-method}
\alias{ps_tibble,XStringSet-method}
\alias{ps_tibble,phyloseq-method}
\title{Coerce phyloseq objects to tibbles (tbl_df objects)}
\usage{
ps_tibble(x, ...)

\S4method{ps_tibble}{otu_table}(x, pivot = TRUE, .name_repair = base::make.unique)

\S4method{ps_tibble}{sample_data}(x, .name_repair = base::make.unique)

\S4method{ps_tibble}{taxonomyTable}(x, .name_repair = base::make.unique)

\S4method{ps_tibble}{XStringSet}(x)

\S4method{ps_tibble}{phyloseq}(x, tax = TRUE, ref = FALSE, .name_repair = base::make.unique)
}
\arguments{
\item{x}{A phyloseq object or component.}

\item{...}{Unused; for methods.}

\item{pivot}{Whether to pivot the data frame to long format.}

\item{.name_repair}{Function to repair names in the case of conflicts.}

\item{tax}{Whether to include taxonomy data.}

\item{ref}{Whether to include reference sequences.}
}
\description{
An S4 generic function that serves the role of \code{tibble::as_tibble()} on
phyloseq and phyloseq-component objects.
}
